LinePlot implementation
